Rename sheet_physics to physics
