+++ package/lib/src/navigation/navigation_sheet.dart
@@ -7,7 +7,7 @@ import 'package:smooth_sheets/src/foundation/keyboard_dismissible.dart';
 import 'package:smooth_sheets/src/foundation/sheet_activity.dart';
 import 'package:smooth_sheets/src/foundation/sheet_controller.dart';
 import 'package:smooth_sheets/src/foundation/sheet_extent.dart';
-import 'package:smooth_sheets/src/foundation/sheet_physics.dart';
+import 'package:smooth_sheets/src/foundation/physics.dart';
 import 'package:smooth_sheets/src/foundation/sheet_status.dart';
 import 'package:smooth_sheets/src/foundation/theme.dart';
 import 'package:smooth_sheets/src/internal/transition_observer.dart';
